requirements.txt
langchain
langchain-groq
langchain-community
langgraph
chromadb
sentence-transformers
python-dotenv
pypdf